september 16 patch (not complete)
several things still need fixing
- EQADDR_MOUSE is broken
- hack detection routines need checking, there were changes to them.....
(this is not safe to run, nothing has been tested yet...)

Requires changes to macroquest repo, money went to 64 bits.
I updated the physics this time, maybe

+++ src/game/PcProfile.cpp
@@ -20,22 +20,22 @@
 
 namespace eqlib {
 
-int BaseProfile::get_Platinum() const
+int64_t BaseProfile::get_Platinum() const
 {
 	return pLocalPC->GetPlatinum();
 }
 
-int BaseProfile::get_Gold() const
+int64_t BaseProfile::get_Gold() const
 {
 	return pLocalPC->GetGold();
 }
 
-int BaseProfile::get_Silver() const
+int64_t BaseProfile::get_Silver() const
 {
 	return pLocalPC->GetSilver();
 }
 
-int BaseProfile::get_Copper() const
+int64_t BaseProfile::get_Copper() const
 {
 	return pLocalPC->GetCopper();
 }
